move files to better locations
