+++ src/main/java/com/fortify/fod/parser/converters/BsiUrlConverter.java
@@ -1,6 +1,7 @@
-package com.fortify.fod.parser;
+package com.fortify.fod.parser.converters;
 
 import com.beust.jcommander.IStringConverter;
+import com.fortify.fod.parser.BsiUrl;
 
 public class BsiUrlConverter implements IStringConverter<BsiUrl> {
     @Override
